Show "system" identifier for winget package count (#104)
* Show missing "winget" identifier.

The choco and scoop packages count both have the package manager identifier written next to it, but winget is missing it. This adds the identifier to keep the output consistent, if users explicitly enable winget in settings.

* "winget" changed to "system"

As the reported number does not correspond to the number of apps installed through winget (but to the number of all installed apps in the system), using "system" seems more appropriate.